Update issue 687
Adding some inline documentation to the new pollinator service value
calculations.

When taking a closer look at the output rasters of the InVEST 2.2 and InVEST 3.0
versions of the pollination model, the outputs are actually pretty similar ... I
must have accidentally been exploring the wrong raster when I made my previous
comment (in a previous commit) about the inaccuracy of the InVEST 3.0 raster.


Former-commit-id: 963702822c0088290016161106c1c2380921d9be