Added functionality to build up shapefile from csv file.
CK gave me a text file with lat/long and other field info.
I converted this text file to a csv file and then created
a function that reads the csv into a dictionary.

I then wrote another function that takes that dictionary and
creates a shapefile from it, using the lat/long for geometries
and the dictionary info for fields


Former-commit-id: d15233ad10ad5a5b6fbbe4801eaa8b82fd04f93c