Update issue 775
This was an issue specifically with the HideableFileEntry class, which was
returning the file field's value as long as the element was 'enabled' (meaning
that the user could interact with the element).  As of this changeset, I've
reimplemented the isEnabled() function of QWidget here in the HideableFileEntry
class.  In this new function, we return False if the element is truly disabled
(meaning that the user cannot interact with the element).  If the element is
not actually disabled in this manner, then we return the check state.


Former-commit-id: 560184feb7a349a925e9442b792daa09a1f8dfbf