Update issue 779
The issue here was that DynamicPrimitive didn't know how to cast the value
of the checkbox, so it automatically tries to cast to a python str.  To fix
this, I added an entry to Registrar.TypeRegistrar for the python boolean type
and I set the checkbox to  have a boolean output by default (unless the user
designates that it should be cast differently, defined in the JSON definition).


Former-commit-id: 50a5945b04e4983ac3daa47f2b8dadfe4e9acbed