Update issue 770
Adding valuation rasters for the current scenario where pixels may or may not be
agricultural.

I decided not to add separate regression rasters for scenarios where pixels may
or may not be agricultural, since the valuation component only cares about ag
pixels in the form of the ag raster, so having two sets of regression rasters
doesn't actually test anything.


Former-commit-id: eddc7a7b40896c8a779db5b9ceb4d9ea24f41d91