Changing the version number to be consisten with distutils's version numbers.  Specifically:
Encoding the version information is an art in itself. Python packages generally adhere to the version format major.minor[.patch][sub]. The major number is 0 for initial, experimental releases of software. It is incremented for releases that represent major milestones in a package. The minor number is incremented when important new features are added to the package. The patch number increments when bug-fix releases are made. Additional trailing version information is sometimes used to indicate sub-releases. These are “a1,a2,...,aN” (for alpha releases, where functionality and API may change), “b1,b2,...,bN” (for beta releases, which only fix bugs) and “pr1,pr2,...,prN” (for final pre-release release testing). Some examples:

0.1.0
the first, experimental release of a package
1.0.1a2
the second alpha release of the first patch version of 1.0

Coding this one as 2.3.0a1


Former-commit-id: d248d0f030a351253bb6f3a3874270defc9b2c65