Update issue 790
Actually, wasn't a problem with vectorize_rasters at all, but rsupply_vol_op was returning out_nodata as a nodata value which was set to -1.0 earlier, even though the vectorize_raster function creates the nodata as 0.0 (correctly).  Changing to have a specific variable in rsupply_vol_op which I'm defaulting to 0.0 and explicitly passing that as the nodata to vectorize_rasters for readability.


Former-commit-id: 78e59bb06a52bff3558989aa7a923343638828b7