Update issue 397
Adding a validation button to self.elements.  This button is disabled until
validation is performed, at which point it becomes enabled.  When enabled, it
displays a green checkmark when validation passes and displays a red x when
validation fails.  This changeset also includes a brighter red x icon ... the
last one was pretty drab.


Former-commit-id: fe9a47bc626760a180d582127fb5665d9d51af32