Update issue 757
As of this changeset, all widgets are checked for a sizeHint.  The widget's
minimum size is set to this sizeHint as long as a sizeHint exists for the
widget.

This should fix issue 757!

note that this implementation does not actually set the minimum size of all
sub-elements of elements themselves, and yet the setMinimumSize functionality
still seems to work well.


Former-commit-id: bb8220d5cd078f148258e9fc509ccc2cbabd2163