Update issue 793
I realized that my old implementation of just CSV file validation was always
passing, regardless of the file actually provided.

Part of the problem here is that DictReader will happily open anything that's in
plain text.  This makes it rather difficult to validate, though there should be
at least some obvious cases when validation should fail.

With this changeset, I've replaced my old implementation with a csv Sniffer
object.  If it honestly can't detect a dialect for the CSV file, it throws a
csv.Error exception.

So at least it fails sometimes :)


Former-commit-id: c609192120551bc0581caa3da681d9e1a06119c5