Update issue 799
Adding some comments and a new default validation scheme for the FileEntry
class.  The problem was that validation of 'type':'exists' was being applied by
default to all instances of FileEntry, even those that actually represented
folders (since type:'folder' and type:'file' both use the FileEntry class).  As
of this changeset, default validation now exists for both the file and folder
types of the FileEntry object.


Former-commit-id: b45b50127a89364e182115561c731d518fc9ea7c