I ended up converting these files to unix coding
I ran into an issue where I was trying to do some 'sed' commands on the files
and was getting some issues where my regex wouldn't work.  It was due to
how the end of the lines were being specified.

Edited sampleWaveDataTest.txt to format the new WW3 text inputs


Former-commit-id: 92a791e3ebc9db08256dbd9a13b122c3795fec95