Fixes issue 815
To avoid perfect alignment of pixels we stochastically perturb input points by +/-1e-6 fraction of the pixel size from a standard deviation.  This way it is nearly unlikely we'll get 3 points that perfectly align along any axis without introducing much error in the result.  Arguably the error is not larger than the error associated with the interpolation scheme.


Former-commit-id: 6ce8dee3a04ea504736e5353e30ea0d85d18a1f4