Update issue 778
Changed how the model treats land and nodata values to be the same.  Since the point-wise inputs sometimes don't define outside the grid, there's little reason to expect an accurate solution there.  Plus the boundary condition of the governing equations is boundary at infinity.  Thus it is reasonable and CORRECT to nodata out the land and nodata values and re-route concentrations around it.


Former-commit-id: baf804f3d25be13546914eed758ede4418d9fa41