Fixing a sneaky bug when validating CSV files
I came across an issue where I was entering a CSV file and it was failing
saying that it needed to be a CSV file.  I got very confused because the
file was indeed a comma separated .csv file.  The issue turned out to be that
the line

csv.Sniffer().sniff(open('csv_test.csv').read(1024))

was only reading in 1024 bytes and my file was larger than that
so it failed to determine a delimeter.  I just deleted the 1024


Former-commit-id: 7c7c6f2ac27aca384207f303a545795757d81f98