Update issue 827
I have solved the issue, and it is so so bogus.
The issue was in the out_nodata value.  When I was in the fractp operation
and looping through the inputs and checking nodata values, if I found a nodata
value I simply returned out_nodata.  HOWEVER, out_nodata was -1 and apparently
it does not like that it is not a float.  So I casted it to a float early on in the
program and it all works fine now.

This seems very odd and crazy.  I haven't looked at vectorized rasters to find
out why an int (-1) would cause everything to flip out.  I'm almost wondering
if by setting the nodata value to an integer if it cast everything else to an
integer making all my results round to 0.  Idk, something weird was happening
for sure.

I'm going to leave this issue open for Rich to close because maybe this is
something we should look into fixing / finding out what really was happening.


Former-commit-id: 9195e86b11a1be29bba4fce3300e40c5c813c342