Update issue 814
Adding some conditional logic to the folder checking class in iui_validator that
uses a new validateAs flag: "mustExist": bool.  If the user does not provide
this attribute, it is assumed to be false.  If "mustExist" == false, the uri is
checked to verify that the uri doesn't already exist as a file on dist. If
"mustExist" = true, the uri is checked to verify that the path provided actually
exists on disk and that the path provided points to a folder.


Former-commit-id: 7af75c11af4f7a86221c7814fe1ad54fc859f151