Update issue 733
Changing format of finfish_aquaculture so that it passes in a reference to the overarching args dictionary, and allows the inner functions to change it, rather than making ff_aqua_args a global variable. When comparing, however, need to only look at the parameters that would be in the individual farm_ops table, rather than the general_ops as well. Otherwise, the sample dictionary will never match the one created.


Former-commit-id: 1234e650e264078f5669afd30310c7f584b2bae3