Update issue 827
Adding a test in vectorize_rasters to ensure that the nodata value type matches the gdal `datatype` that's passed in.  Otherwise we can end up setting a nodata value that might not match the datatype.

Note we still need to ensure that the implementation in 'op' matches all the correct types, no way to do this within the funciton, we just need to be careful as programmers.


Former-commit-id: 4ca6a7de5323136d602ac8f3ac8f4062c53704c7