Update issue 855
Wrote a new function get_area_of_polygons that I hope will take the place of
create_area_raster and be less of a memory hog.  This function simply
returns a dictionary that matches the shapefiles id's with its area.

I'm thinking that I can then use this, a mask, and vectorize_rasters
to create an area raster


Former-commit-id: 224f3ddecc4b843bcb7fa95710ac4c7042e1bc06