Fixes issue 885
Adding a conditional that checks to see if this qwidget is invisible.  If so, we
should enable it.  This somehow prevents the widget from being shown all the
time, which was what was happening before ... the flicker was the MessageArea's
error colors being set and then being shown right then, despite the fact that
the parent widget had not yet been shown.


Former-commit-id: 00259dbeb5d4440889b3c5f471ddb59026db9306