Update issue 897
Adding a setState call to the check_validation_error callback so that dependent
elements are toggled based on the validation result.

Also, base_widgets.TableHandler now checks the error status of the enabling
element to see if it should be enabled.

Also, fixing an issue with fileio.AbstractTableHandler.update() where the new
table was not being opened for reading before the internal table data structure
was being rebuilt.  My bad.


Former-commit-id: ea8e225a92bb795e25064da5c61ec1f2ad4a1b7d