Update issue 896
I was reusing variable names for rasters and wonder if maybe this could be an
issue related to file permissions and tiff errors when running on Windows.

I changed the names so that it is much clearer about what rasters are what.

The only other thing I've come across is that maybe it is a projection issue
or that change_shape_projection() is doing something interesting. Because the
only difference between having an AOI and not having AOI is that their is no
proper projections without an AOI and change_shape_projection() is not called
without an AOI.


Former-commit-id: be6adbc7ba9a3d959311736b130f8389136daba2