Update issue 723
Skipping the default sample input for because the default raster files are too
large to add to the repository through mercurial.

For Hydropower it seems that we don't clip some of the rasters to the area of
interest or water sheds so they end up being quite large.  We are also using
very fine 30 x 30 resolution from the lulc.


Former-commit-id: 511020971236c6fa1b60d5cac6c65f7a080f000c