Update issue 917
The reason I was struggling to implement the gaussian filter was because the
input matrix I was providing had a numpy array type of integer (I believe) and
so when it was running through the gaussian filter and returning, it was
returning an integer type.  In order to get it to be a float I cast the input
matrix to be of type float, and that fixed my issues.


Former-commit-id: c4e206a5344e2d73d6f931adad8ad474a75b63a9