Fixes issue 916
I created a helper function that takes a uri which includes the directory path
as well as the base file name.  The function then has a pre-defined list of
possible suffixes (a directory, .tif) and attempts to open the uri with
gdal.Open by concatenating a suffix with the uri.  If the returned dataset is
not None, it breaks and returns that dataset.

In the instance where each dataset returns None the function will return None
and then will be handled properly back at the URI level.


Former-commit-id: c3d98647cd2d3f95c0397db65ff6a9a03aec709b