Update issue 932
I added a logging statement for what is being passed in as sigma to the
gaussian_filter function.  I do not know what the proper sigma should be at this
point.  Plugging in the -2.99573/max_cell_dist seems to yeild such a low result
(< .10) that it does not affect the dataset.

So I am lead to believe that this may be wrong.  Or it may work in the current
implementation because they are using a resolution of 1000 x 1000 and I am using
30 x 30.  If I used something like a 1000 then my sigma would be closer to .4
which would have some affect using the gaussian_filter


Former-commit-id: e82e19001b5344cd0ee085fd60c44ade21c26757