Update issue 935
Changed some of the naming conventions in uri handler for the keys of
dictionaries.  Also now store any lulc_cover in a dictionary which is past into
core.

Made edits to core to try and use this new setup.  Now loop over all landuse
covers and try to perform proper actions.  Added docstring better describing
landuse_dict that is being past in.


Former-commit-id: ffb0e02d5779764157455b8df057ca59d713e78e