Fixes issue 948
Verified that flow accumulation is working with only a very small roundoff error, as might be expected from dinf passing flow from cell to cell using 32 bit floating points.  Explected hand calculated value was 9604, get 9603.89 after multiplying across 9604 cells.  This is within range of a roundoff error, but suggests that a linear system may be a better mechanism to account for flow accumulation.


Former-commit-id: 969c2281f93c752c970794ae980a2b5c9315cb47