Fixing an issue that Greg V brought to attention with wave energy
There was an issue that Rich fixed in raster_utils.py with
create_raster_from_vector_extents.

I also had to make a change in the core and reset the iterator head for the
shape file being passed in. In the future, and a good refactoring project, would
be best to implement iterating over features and layers the way Rich came up
with.


Former-commit-id: 40103fcff34e4331718a9a4f631841cf6a7a28c0