Update issue 940
Previously I was using DynamicGroup's createElements function to actually create
the element of choice.  Now I directly create a local element and try to get the
file name desired by the user.  If the user selects a file in the dialog, the
file widget is added to the visible widgets.  Otherwise, the element is ignored
and eventually goes out of scope (where it is eventually deleted).


Former-commit-id: 115aa92463d92a475a4c155c301402c49babfffc