Update issue 860
Wrote the function to create the layers dictionary. It will contain the ID numbers as keys to the dictionary, and have values of lists [inter-activity weight, buffer].

This will completely bypass the need for text-based comparisons of shapefile names and names from the table, since they can't be counted on to be the same. Instead, when examining the shapefile names, we will parse the filenames to look for the integer in them (since there should only be one) and then use that to pluck the correct information from the dictionary.


Former-commit-id: 10af7c2a80d51ae874c5e954aa54cad18e8f9f8c