Update issue 860
Changed the layer type to be an ogr.wkbPolygon, as opposed to a wkbPoint file, since it won't let you put full polygons on a shapefile designated as entirely "point."

In addition, had written out closeRing(), instead of closeRings(), which was causing the program to error. Fixed that.

Need to changed the order in which the points are added so that it goes in a square shape, instead of crossing over itself and creating an hourglass.


Former-commit-id: c89080d471a9a846add3ae4395f93d0dd5c82762