Update issue 933
Finished structure for calculating the rarity output.  I added basic comments
and doc strings to the code and operations.  Rarity is now figured by mapping
the rarity index to the current / future land cover maps based on lulc codes.

I still don't think that the map is trimming right to be the same size as the
base raster is.  Instead of writing nodata values it is filling it in with
zeros.  Something to figure out.


Former-commit-id: ffdea79b77327a82caba6c15ecf74ac7ed4922de