Update issue 933
Trying to figure out why when I can't make a new land cover and set the nodata
values properly.  I have a base raster which has a section of nodata values and
a current raster which is filled in.  I basically want a new raster that is the
current raster masked to the base rasters nodata values.  But it's not working,
where nodata values should be, I get 255 (which is the nodata value of those
rasters).  I set the nodata value and return -1.0


Former-commit-id: d9661713c84d37214d4a469650b7b8871fb75489