Update issue 933
Ok, so the issue with the nodata values comes from a similar issue that we have
had in the past with passing in different numpy types to vectorize.  This issue
was my two input numpy types were 8 bit integers and I was trying to return
either a float nodata value or a 8 bit integer onto a float32 numpy array.  This
caused all sorts of weird issues.  But by casting my 8 bit integer result to a
float it seems to act properly.  So that is what I do.

IF ANYONE READS THIS TAKE NOTE!  IT COULD HAPPEN TO YOU. IF SOMETHING FUNKY IS
HAPPENING IN YOUR VECTORIZE OPERATION IT MAY BE A TYPE ISSUE.


Former-commit-id: 43e0d91b1141a9bb63644b51bd712ac5cf04ec96