Update issue 933
There was an issue with setting the nodata value for the final output raster of
mapping the lulc index dictionary to rarity.  I was using a very small number
close to zero for nodata which ended up making any zero value be nodata.  I now
use a very large negative number for my nodata value.

I also changed the output type of the 'new_cover' which is the cur/proj cover
masked with the baseline cover. It is now Int32, because the lulc codes are all
integers.


Former-commit-id: 39641f68a47eb28cc61c171f6f05140599c8eb4b