Update issue 968
Added a check in uri handler that builds up a list of tupples that represents the resolution (x,y) for each land cover provided.  If the resolutions do not match, we print an error and set a variable to exit.

Should I be using sys.exit to just quit the model?


Former-commit-id: 76478feda19ed86f5b16bb359ff3dacf03e903ca