Update issue 968
Removing resolution checking as Rich and I think that it should no longer be
necessary to have the same resolutions.

Actually first let me say that I have been mis-using resolution to mean pixel size.
So we no longer need to have the rasters be the same pixel size.  When computing
rarity, equation 12.5 should reflect the ratio index in terms of pixel areas

ALSO: now raise an exception if projections of land covers fail.  this should be
in issue 963

Former-commit-id: fa444d54775c86bdd614befd1765c5841f7c6dd9