Update issue 964
I refactored the function 'raster_from_table_values'. Changes include:
- new name : 'raster_from_dict'
- two new arguments, 'raise_error' and 'error_message'
- nodata is now required as an input and has no default

The function now raises an exception if wanted to do so. This solves the issue
where we may have a lulc type on cur/fut/base that is not in the sensitivity
dictionary.

I still need to create a few test cases to make sure this works properly.


Former-commit-id: e72a39e700675b824030771d6480f77334d4b555