Update issue 860
Files are now correctly retrieved and added to a list. However, get_raster_sum is broken. The error log claims that raster_utils.vectorize_rasters is taking in an invalid number of arguments. However, despite matching the number/type of arguments that is in the raster utils docstring, it refuses to continue.

Have tried with and without the aoi parameter. (Did an example without, and that worked, so tried it with 5 paramaters too.)

>raster_utils.vectorize_rasters(raster_files, get_raster_sum, aoi=None,
                                   raster_out_uri = activities_uri,
                                   datatype = gdal.GDT_Int32, nodata = aoi_nodata)


Former-commit-id: b61ae5c1107d9d866ba1c0660051a2b880481e81