Update issue 786
Wrote a test case for create_rat function in raster_utils.  This test function
test writing a new RAT to a band and then reading it and checking its values.
The next test should try and edit a RAT that is already assigned to the band.

I also added logging to this raster_utils_test and osr


Former-commit-id: 61061d98f7abd84e6dd03029b620ed618572feab