Update issue 786
I now sort the keys as a numpy array so that when adding them they are
guaranteed to be in order.  I'm thinking that it would be useful if the LULC
codes being our keys were in ascending in the rows of the RAT.  I think this
is an ok rule in general.  When creating a RAT sort by the keys.  Also added a
line that increments row_count which I wasn't doing.


Former-commit-id: 372b4971cea1193ef70630ce57fc2123407d14bd