Update issue 860
Got the rasters to line up so that there is an overlap after all. Before, I was improperly using the values that I was getting from the current AOI, and creating a grid that was at opposite coordinates from the actual AOI. Now, the rasters that we are creating are in the proper place.


Former-commit-id: dcf0a91e682db88d6bbbba78ca40822890a54429