Update issue 786
Created a second test that sends in a dataset that already has a RAT. It
properly adds the next two columns and values properly.

The function is currently really restricted.  You should be able to add as many
columns as you want with however many values.  Should decide how versatile and
complicated we want this function to be.

Should also test instances where the number of rows being written to an
existing RAT is different than the current number.


Former-commit-id: ba1d2dc932a96d4d7e30b2981da116065f4d6916