Update issue 957
Now, in iui_validator when a strange exception is caught in a Checker class
(like when input somehow causes validation to crash), a Warning is raised and
caught so that a tuple can be returned to the user indicating the error message
and the state of the error (whether warning or actually an error).  Also added a
get_error() function to assemble this tuple.


Former-commit-id: 3e66c549f881f7990637f23389c86760b5fa6701