Created the full json for the OA model. Then updated the args docstring to reflect that in the overlap_analysis.py file, and made it so that the inter weights dictiobnary is an option pass to the core module. When the argument for intra is added, need to make sure I only pass it along if do_intra is true.
I am making a change to the model in that both the layers table and the attribute kin the shapefile are both optional, since the model does not need to be run with either inter-activity weights nor intra-activity weights. Since the files are being sorted into a dictionary by name now, rather than the ID number, there should be no confusion when the CSV for inter is loaded, since it will be able to easily reference who it belongs to, as their names will be the same.


Former-commit-id: 57786c10c96503a8ae503026312d9cead964c437