Update issue 860
Created a skeleton for the weight raster created function. This will output a raster file that takes into account both the inter-activity weights and intra-activity weights.

Question: If we want neither inter-activity weights nor intra-activity weights, do we want to produce this output at all? I suspect that it would look quite like the other raster that we're producing, and would certainly be adding nothing.

Former-commit-id: 1bfa1250c586126bdac6a742442cd7fa99daa0e0