Udpate issue 860
Need to create temporary weighted rasters where the burn value for the raster files is intra_activuty weight for each pixel within a polygon on the shapefile. Added the ability to create a set of temporary rasters based on what's in the layers dictionary.

Additionally, changed some variable names around so that the correct values will get passed when do_inter and do_intra are True, but otherwise will just pass None values so that they can be ignored.


Former-commit-id: 616b083305461308b1d482f9751e89eb93f65780