Update issue 860
SO MANY CHANGES.

We are now preperly able to rasterize the layers with their attribute weights. However, it is currently still being done as a text input. This is fine, though should be aware that if you try to rasterize using a non-existent attribute, it will create a raster, just blank.

Within the create_weighted_rasters function, we are also able to create a dictionary of the max intra-activity weights for a given layer. Unfortunately, in order to do this, we are currently looping over all of the features in all of the layers. While it is n^2 time, it's not so bad, since the datasets tend to be small and in manageable numbers. We throw the

Additionally, we are getting the max inter activity weight for a given layer by just doing a max() of the set of inter_activuty_dict values.


Former-commit-id: 91373b879851ef619f767315205dd412fe7fe817