Update issue 935
In the current model if a threat/density raster is not provided or found for a
landcover it skips to the next landcover. I was just ignoring this threat and
carrying on.  I added a check in the code, where if we come to a threat raster
that is None, we skip to the next landcover.


Former-commit-id: 12b552cca12dd4dfe237c9ec7a0c5c155528b515