Update issue 935
Since I was only skipping threats if their density raster was not found I was
checking for the case where no density rasters where found so we didnt try and
compute the quality raster. Now that we break if even one density raster is not
found this check is no longer needed.


Former-commit-id: 9c101af5f06cb510435e77fc3a88a2443fb5f223