Update issue 860
Re-writing the way in which I create the weighted rasters. I need to know the order in which the rasters exist in my list. So instead of creating them with the bigger function of create_weighted_rasters, I am instead moving that process over to make_indiv_weight_rasters.

This function will return a tuple of lists. The first lists will be the same as the old return- a list conatining the open raster datasets that are produced when rasterized based on the aoi. The second will be a list simply containing the filenames, in the same order as they were added to the first list. This will allow me to know what order the pixels are in when I iterate through them for the final weighting calculation. This name can be used to index into all of my other dictionaries containing the inter weighting information.


Former-commit-id: 8d0b57edf399880c8b6fadb89afe51384fd4ae88