Update issue 860
Ran changes from last time. The make_indiv_weight_raster function now works, and does indeed create the files. However, when they are examined at the end of the model run, there is no data showing up. If you print out the attribute, it does show that the intra_name variable does exist, it just doesn't seem to be burning things correctly to the rasters. I suspect this is an issue with the "option = " field in the RasterizeLayer call.

Perhaps something to check for tomorrow.


Former-commit-id: cae76c49c356609ad06eca4cbd5aef5e2dccbdc7