Update issue 860
Fixed the make_indiv_weight_rasters function so that it works properly. Should now create rasters where the burn value is the selected attribute. However, need to be careful in the future that there are NO SPCAES when using the element = ['ATTRIBUTE=%s' variable] notation method. If there are spaces, the whole thing breaks.

Now trying to put together the entire equation to spit out something.


Former-commit-id: e46a207d33f9956619ea529ceb527cb2afa31f89