Update issue 860
Added a separate function for adding the intra-weight rasters. Haven't yet tried to run.

To consider: The two functions could be combined with the use of some more if statements within the pixel combination functions. Would this be desirable over separating them with if statements outside? Note: this would not remove the need to if statements within the make_weighted_raster function, but it would mean that we could pass the same function to both cases, just using different lists of rasters.


Former-commit-id: 718d7dfadc7b05f23aae527a96325cc0845d7737