Update issue 860
Created textfile output. Should be noted that like the textfile output for finfish aquaculture, this explicitly names the parameters that are being written into the file.

It should be noted that some things that the user is passing in within the IUI (the location of the CSV, for one) are not able to be written to the textfile, since they are in dictionary form by the time they get to the OA core module. Is this something that we really want to write back? If so, I can just add an extra argument into the args dictionary being passed.


Former-commit-id: 263fb97f9cd3f05c0df1035a73d08321557c3992