Update issue 962
Pulls out the files completed by the test fun. Need to compare them against something. Is there a way to rasterize a vector by hand in qgis? That way I could create my 1000x1000 pixel blocks, and just set them to whatever I calculate the output of the frequency or the weighting should be.


Former-commit-id: aa76d676ad72829dcbfd5cc1855c969b8d063509