New issue
Summary: Splitting Managament Zones and Gridded Rasters in OA

Initializing new files for splitting MZ and the original OA. MZ will simply do an activity count in each of the polygons, and output a modified shapefile as a function. The other overlap analysis, which has yet to be named, will perform the original functions of overlap analysis.


Former-commit-id: cfa7271ac28d77dd7a9ddf79b8b821f19750918c