Update issue 1087
No wind data point is detected within the area of interest (BarkClay). This is strange, since this aoi is the one mentioned in the model's documentation. From visual inspection, some points seem to be in the AOI. When I try to rasterize the WW3 dataset within the AOI, ogr gives me an error because I seem to try to burn values outside of some permissible limit...


Former-commit-id: 700f3efad62847439c525d50e8a3b6e272935d12