Update issue 935
I had an error where if a threat / density raster was not found for a landcover
it was not properly exiting and moving onto the next land cover. I have nested
for loops and now I break out of the first which iterates through the threat
rasters and continue on the second which iterates through the landcovers.


Former-commit-id: 68f5ce8a421850e97df732a46ea7a6b8d0074325