Update issue 844
Fixed an issue where I wanted to raise an error if a lulc code in the
landcover was not found in the sensitivity table when mapping sensitivity data.
There was an issue where a nodata value would from the lulc raster would pop up
and try to be found in a dictionary, but it obviously wasn't in there, so it
raised an error, when in fact this should just return a nodata value.


Former-commit-id: f7c060f1e79e20bf9798cc4b27c76099c390c9ba