Update issue 1144
Split the unweighted function to add parallelism into the oa_core module. Since that allows the function names to be nearly the same with weight/unweighted, I am not updating teh function names. Now the execute will only run unweighted, then check to see if it needs to run the weighted side.


Former-commit-id: 83da8c0bd01873d31965fd588bc12d2d71b16cdd