Update issue 1147
Removed commented out code and am committing to computing degradation in one
step. Added a ton of comments.

Also found out that when a list is past into the operation of a
vectorized_rasters call, it is a tuple and not list.

After testing I have come to the conclusion that map_raster_from_dict_values is
what is taking the most time and will be what I look at next


Former-commit-id: b269bddbe237e0b56d6f5a85ab1d4ee04ee3c3e7