Update issue 1146
Fixed a tricky bug.  When rasterizing attribute fields I was running into a problem when I filled the band with 0 then rasterized a field value of 0 over everything.  Instead made nodata in mask band 255 and explictly tested for it depending on if there was a field to test for or nothing at all.


Former-commit-id: 5dc1cfb6d8c915e615103e14ff3152e7ec77e5d1