Update issue 1147
Re-working how I handle access_shape. Now if an access_shape is provided I
create a raster from it's values, if it is not provided, I create a raster
filled with 1's. This way when I call total_degradation I can just pass it in
regardless and don't have to worry about any if/else statements or checking to
see if it exists


Former-commit-id: 0b70a264229a3f1406d227219fac31a7af565fdc