Update issue 1177
created a first draft of using the min/max values, and aoi to create a depth
mask. I used vectorized_rasters to clip the bathymetry raster, and only return
valid data if the value was in the range of the min and max depths, otherwise
return nodata.

I do some multiplying by -1 to get the min max into the same format the
bathymetry values are in, probably should do this at the URI level


Former-commit-id: eb953fde7d4719c47b4d9adeedc1f20ad5b304df