Update issue 1178
I was running into some covolution issues, like the scenario where I have land
values that lie on the outside edge of the raster. How can we tell if it is
coastline or not? So what I have decided to do is to create another raster that
is 2 rows / columns greater than the raster created from the AOI. This way I can
set the outter perimeter with nodata values and not have to worry about that
edge issue ( I think ).

Created a second raster 2 rows / col bigger than AOI raster. Wrote the land
matrix with all burned values onto this new raster with an offset of 1 for both
x and y.

Also properly set nodata values for the land dataset. Before I was just burning
land over the AOI without resetting the nodata values to reflect that of the AOI


Former-commit-id: b7333148ca129fc31d7e7e918a79ab4af81edba0