Update issue 1178
Created the boundary raster, where the land / ocean values are buffered with a
row / col of nodata values. Then used this to do convolution magic against with
a kernel. I reset the nodata values and now have a matrix that should be my
proper shoreline. The only question is if I want to trim the extra row and
column off (I don't think it's necessary) and write it out to a proper raster.


Former-commit-id: 3eab02bea2dd2773071df9e507322237fd3d0641