Update issue 1188
Setting up for computing distances with gaussian fitler. Copied some code from
the biodiversity model and commented it out. Will use this as a starting point
for writing out what I will need.

I am not sure what I should be using for the sigma value in the guassian blur so
I derived sigma like we did for biodiversity, except this time for linear decay.
I'm not sure if this is right. sigma = sqrt(distance_in_pixels / 2)


Former-commit-id: 448bb208efbe3132b76994be49858c8a9a9e0b9b