Update issue 1191
Implemented the test for identical projections, which uses 3 new functions, 2 that extract the projection from a raster or a shapefile, and one that compares projections to see if they are identical. Implementing the functionality that gets the size of a pixel at a specific location. However, since the shapefile is an aoi, the location is a polygon, not a point, and I can't get x and y coordinates from this polygon.


Former-commit-id: 339f271a8b71556a3d794b28fc8714be3c35d0fa