Update issue 1204
I created two functions in wind energy biophysical that deal with reprojecting
and clipping given two different circumstances.  The first function is and idea
for a general situation that may be useful to have in raster_utils. It deals
with the instance where you have an unprojected dataset and a projected
datasource that is the AOI. You want to clip and reproject the dataset to the
datasource. The issue is if the dataset is a DEM or another large raster file.
You really want to clip the raster before reprojecting and in some instances you
have to because a global dataset is hard to reproject because the X values end
up being the same so when you get the range for X width it is 0, which is
illegal.  Thus the idea is to back project the AOI datasource to the projection
of the dataset (which is probably unprojected WGS84), then clip, then reproject
to the projection of the AOI. That is what the first function does

The second function is really just a space to collect my thoughts on what should
be done for the wind energy model. I have an AOI, Land Polygon and DEM. Since I
have multipled shapefiles I have the luxury of just clipping the DEM from the
AOI and reprojecting to the Land Polygon. However, what if the AOI is also
projected in Meters? Should the users guide state that the AOI should be in
WGS84 or should we account for the fact that it may be in meters? Descisions,
decisions, decisions...

I think what I will do is check to see if the AOI is projected, if so, I will
check to see that it has the same projection as the Lany Polygon (if provided).
If the AOI is projected and the Land Polygon is not provided then I don't think
it hurts to call the first function listed above and just do everything in the
projection.

Wow this is long, but I am feeling very hyped and focused from coffee this
morning... what was in the coffee? Or it could be these intense amazing FILM
Scores that are getting be hyped! Or the fact that I defeated the DRAGONS. If
you're still reading this you seriously need to get a life and find something
better to do, don't you have your own work?


Former-commit-id: af285c1f1c1760d23dca0fdf1ef7eb9236996ce2