Update issue 1204
It has been decided that the AOI must be projected, and that the other inputs
can be whatever. This means that all other inputs are to be reprojected to the
AOI projection. This has been implimented.

A check_datasource_projection function has been added to make sure that the AOI
is projected properly.

All dataset/datasource inputs have thus been clipped and reprojected where
necessary.


Former-commit-id: faf610d03efea53a34039a10d844e1733fcdf5f2