Update issue 1206
Removed the passing of the decay param from the make_weighted_raster functgion. Instead, since we are going to be pre-creating the raster to be passed to the function, it will not need to know about decay. It will only need to use the result of that decay applied to the raster.


Former-commit-id: 09b407e92f8ce3fe4b5e24d61c6538e90d33c550