Update issue 1217
Constructed a harvest_wind_energy function to be called from integrate quad.
Computes the first integration of the equation.

I also set up a loop to go through all of the points and compute the harvest
wind energy value.

Set up the known variables needed from the UI

This is a rough layout to get things functional. A lot of efficiency to work on


Former-commit-id: eaab563815db29b5aef2f4ff4bb60e4969131ca1